moves generation debugging and impl

+++ util/Bitboard/bbHelperFunc.py
@@ -6,4 +6,4 @@ def is_set(bb, field):
     return (to_bitboard(field) & bb) != EMPTY_BB
 
 def to_bitboard(index): # returns the piece in a bb
-    return np.uint64(1) << np.uint64(index)
+    return np.uint64(1) << np.uint8(index)
